+++ src/main/java/xyz/nikitacartes/easyauth/storage/database/DbApi.java
@@ -62,15 +62,17 @@ public interface DbApi {
      * Deletes data for the provided username.
      *
      * @param username username of player to delete data for
+     * @return true if player data was deleted, otherwise false
      */
-    void deleteUserData(String username);
+    boolean deleteUserData(String username);
 
     /**
      * Updates player's data.
      *
      * @param data data to put inside database
+     * @return true if player data was updated, otherwise false
      */
-    void updateUserData(PlayerEntryV1 data);
+    boolean updateUserData(PlayerEntryV1 data);
 
     /**
      * Get all data from DbApi.