+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Incompatibility with QuickBackupMulti mod [#232](https://github.com/NikitaCartes/EasyAuth/issues/232)
 - MongoDB connection [#247](https://github.com/NikitaCartes/EasyAuth/issues/247), thanks to @MemoKing34
 - Server crash during config reload [#248](https://github.com/NikitaCartes/EasyAuth/issues/248)
+- Incorrect messages when deleting/updating non-existing account
+- Sometimes players are not deleted properly from the database
 
 #### Changes
 - Update Chinese translation [#233](https://github.com/NikitaCartes/EasyAuth/issues/233), thanks to @ShadyLeaf